Додав валідацію
