+++ controllers/contacts.js
@@ -1,86 +1,50 @@
-const Joi = require('joi');
-
 const contacts = require('../models/contacts');
 
-const { HttpError } = require('../helpers');
-
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.number().required(),
-});
+const { HttpError, ctrlWrapper } = require('../helpers');
 
 const getContacts = async (req, res, next) => {
-  try {
-    const result = await contacts.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
+  const result = await contacts.listContacts();
+  res.json(result);
 };
 
 const findContactById = async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404, 'Not found');
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
+  const { contactId } = req.params;
+  const result = await contacts.getContactById(contactId);
+  if (!result) {
+    throw HttpError(404, 'Not found');
   }
+  res.json(result);
 };
 
 const addContact = async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const result = await contacts.addContact(req.body);
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
+  const result = await contacts.addContact(req.body);
+  res.json(result);
 };
 
 const editContactById = async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { contactId } = req.params;
-    const result = await contacts.updateById(contactId, req.body);
-    if (!result) {
-      throw HttpError(404, 'Not found');
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
+  const { contactId } = req.params;
+  const result = await contacts.updateById(contactId, req.body);
+  if (!result) {
+    throw HttpError(404, 'Not found');
   }
+  res.json(result);
 };
 
 const deleteContactById = async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.removeContact(contactId);
-    if (!result) {
-      throw HttpError(404, 'Not found');
-    }
-    res.json({
-      message: 'contact deleted',
-    });
-  } catch (error) {
-    next(error);
+  const { contactId } = req.params;
+  const result = await contacts.removeContact(contactId);
+  if (!result) {
+    throw HttpError(404, 'Not found');
   }
+  res.json({
+    message: 'contact deleted',
+  });
 };
 
 module.exports = {
-  getContacts,
-  findContactById,
-  addContact,
-  editContactById,
-  deleteContactById,
+  getContacts: ctrlWrapper(getContacts),
+  findContactById: ctrlWrapper(findContactById),
+  addContact: ctrlWrapper(addContact),
+  editContactById: ctrlWrapper(editContactById),
+  deleteContactById: ctrlWrapper(deleteContactById),
 };
