+++ routes/api/contacts.js
@@ -2,15 +2,19 @@ const express = require('express');
 
 const ctrl = require('../../controllers/contacts');
 
+const { validateBody } = require('../../middlewares');
+
+const schemas = require('../../schemas/contacts');
+
 const router = express.Router();
 
 router.get('/', ctrl.getContacts);
 
 router.get('/:contactId', ctrl.findContactById);
 
-router.post('/', ctrl.addContact);
+router.post('/', validateBody(schemas), ctrl.addContact);
 
-router.put('/:contactId', ctrl.editContactById);
+router.put('/:contactId', validateBody(schemas), ctrl.editContactById);
 
 router.delete('/:contactId', ctrl.deleteContactById);
